+++ runtime/System/SByte.cs
@@ -78,7 +78,8 @@ public struct SByte : IComparable, IFormattable
 					else
 					{
 						return NumberFormatter.FormatFixedPoint
-									((ulong)(-value__), 0, 0, true, format,
+									((ulong)(uint)(-value__), 0, 0,
+									 true, format,
 									 NumberFormatInfo.GetInstance(provider));
 					}
 				}
